show confirmation dialogs for canceling or confirming jobs
